[ASS01] Fix warnings

+++ Source/Library/Game/Game.cpp
@@ -77,7 +77,7 @@ namespace library {
 				elapsedMicroseconds.QuadPart = endingTicks.QuadPart - startingTicks.QuadPart;
 				elapsedMicroseconds.QuadPart *= 1000000;
 				elapsedMicroseconds.QuadPart /= frequency.QuadPart;
-				float deltaTime = elapsedMicroseconds.QuadPart / 1000000.0f;
+				float deltaTime = (float)elapsedMicroseconds.QuadPart / 1000000.0f;
 				// Update game
 				m_renderer->HandleInput(m_mainWindow->GetDirections(), m_mainWindow->GetMouseRelativeMovement(), deltaTime);
 				m_mainWindow->ResetMouseMovement();
